requirements.txt
langchain-google-genai
streamlit
langchain
python-dotenv